requirements.txt
async-generator==1.10
attrs==19.3.0
backcall==0.1.0
bleach==3.1.1
blessings==1.7
bpython==0.18
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
curtsies==0.3.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
docopt==0.6.2
entrypoints==0.3
Flask==1.1.1
greenlet==0.4.15
grip==4.5.2
idna==2.9
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
json5==0.9.2
jsonschema==3.2.0
jupyter-client==6.0.0
jupyter-core==4.6.3
jupyter-server==0.1.1
jupyterlab==2.0.1
jupyterlab-pygments==0.1.0
jupyterlab-server==1.0.7
kiwisolver==1.1.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.0
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.18.2
paho-mqtt==1.5.0
pandas==1.0.2
pandocfilters==1.4.2
parso==0.6.2
path-and-address==2.0.1
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
prometheus-client==0.7.1
prompt-toolkit==3.0.4
ptyprocess==0.6.0
Pygments==2.6.1
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
pytz==2019.3
pyzmq==19.0.0
requests==2.23.0
Send2Trash==1.5.0
six==1.14.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.8
voila==0.1.21
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
